requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
amqp==1.4.9               # via kombu
anyjson==0.3.3            # via kombu
appdirs==1.4.4            # via black, virtualenv
asgiref==3.2.7            # via django
attrs==19.3.0             # via black
autoflake==1.3.1          # via -r requirements.in
beautifulsoup4==4.9.1     # via -r requirements.in
billiard==3.3.0.23        # via celery
black==19.10b0            # via -r requirements.in
celery==3.1.26.post2      # via django-celery
click==7.1.2              # via black
distlib==0.3.0            # via virtualenv
django-celery==3.3.1      # via -r requirements.in
django==3.0.4             # via -r requirements.in, django-celery, djangorestframework
djangorestframework==3.11.0  # via -r requirements.in
envparse==0.2.0           # via -r requirements.in
filelock==3.0.12          # via tox, virtualenv
invoke==1.4.1             # via -r requirements.in
isort==4.3.21             # via -r requirements.in
kombu==3.0.37             # via celery
lxml==4.5.1               # via -r requirements.in
packaging==20.4           # via tox
pathspec==0.8.0           # via black
pluggy==0.13.1            # via tox
psycopg2-binary==2.8.4    # via -r requirements.in
py==1.8.1                 # via tox
pyflakes==2.2.0           # via autoflake
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via celery, django
regex==2020.5.14          # via black
six==1.15.0               # via packaging, tox, virtualenv
soupsieve==2.0.1          # via beautifulsoup4
sqlparse==0.3.1           # via django
toml==0.10.1              # via black, tox
tox==3.14.5               # via -r requirements.in
typed-ast==1.4.1          # via black
uwsgi==2.0.18             # via -r requirements.in
virtualenv==20.0.21       # via tox
